==> scripts/06_create_seasonal_networks.Rmd <==
---
title: "Create Seasonal Participation Networks"
author: "E. Fuller, M. Fisher"
date: "Written Mar. 8, 2019. Last Run `r Sys.Date()`"
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
  pdf_document:
    highlight: haddock
    number_sections: yes
    toc: yes
    toc_depth: '3'
geometry: margin=1in
subtitle: Preparation for network analysis in Fisher et al.
fontsize: 11pt
---

# Description

Creates a seasonal fisheries particiation network (igraph object) for each port group, for every crab year. Then calculates summary statistics for each network.

Section 3 allows you to calculate network metrics from existing igraph objects. Set Section 2 code chunks to `eval=FALSE` and Section 3 code chunks to `eval=TRUE` to run Section 3. 


<br>
```{r "setup", include=FALSE}
if(!require("here")) {install.packages("here")}
library(here)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())

## start time for full script
script_start_time <- Sys.time()
```
<br>

This script requires the following packages. 
```{r packages, message=FALSE, warning=FALSE}
if(!require("tidyverse")) {install.packages("tidyverse")}
if(!require("lubridate")) {install.packages("lubridate")}
if(!require("ggplot2")) {install.packages("ggplot2")}
if(!require("igraph")) {install.packages("igraph")}
```
<br>

And calls the following functions:
```{r}
source("R/generate_participation_network.R")
source("R/color_network_nodes.R")
source("R/network_stats_functions.R")
source("R/calc_network_stats.R")
```
<br>

# User Inputs 

Select your directories.
```{r}
## location of fish ticket with metiers assigned, metier key for each port group
indir = 'data/input/metiers'

## output directory for igraph object
outdir = 'data/networks/participation'

## output directory for network viz
pngdir = 'data/networks/participation/plots'

## output file (including directory) for stat summaries
statdir = 'results/statistics'
```
<br>

Identify the crab years and port groups that you would like to produce networks for. The port groups vector should include only those port groups which are present in the single data file produced with script 05. 
```{r}
## (crab) years
years <- seq(2008,2017)

## port groups
myports <- c("CCA", "ERA", "BGA", "BDA", "SFA", "MRA", "MNA")
```
<br>

Some Dungeness crab landings may be recorded prior to the official opening date of the season as a part of domoic acid testing. We remove these landings because we are interested in flows of fishers between fisheries as a result of within-season activity.
```{r}
rm_crab = TRUE
```
<br>

For confidentiality, three or more vessels much be participating in a given fishery for that fishery to be included in the networks. To avoid inclusion of rare or minimal fishing activity, a given fishery must contribute to at least 10% of a vessel's seasonal revenue to be included in the network data. 
Both cutoff values can be increased if more stringent fishery inclusion criteria are preferred, and the contribution cutoff may be decreased (i.e. for summaries over shorter temporal / smaller spatial scales).
```{r}
vessel_cutoff <- 3
contr_cutoff <- 0.10
```
<br>


Set these objects according to your user inputs from Scripts 1 and 3
```{r}
## the value of `k` for the k-nearest neighbor function
k <- 1

## the reference years first run through infomap
ref_years <- c(2010,2011)
```
<br>


# 1: Read in data

Read in the landings data, from the file containing all fish tickets across port groups, within a single crab year.
```{r}
for(y in years){
  if(y %in% ref_years){
    tmptix <- read.csv(here::here(indir, paste0("FishTix_crab", y, "_metiers.csv")), stringsAsFactors = FALSE) %>%
    filter(pcgroup %in% myports)
  } else{
  tmptix <- read.csv(here::here(indir, paste0("FishTix_crab", y, "_metiers_k1_recoded.csv")), stringsAsFactors = FALSE) %>%
    filter(pcgroup %in% myports)
  }
  if(exists('fishtix')){
    fishtix <- rbind(fishtix, tmptix)
  } else{ fishtix <- tmptix }
}
rm(tmptix)
```
<br>

Get the opening dates associated with each port group, for each crab year. 
```{r}
dates_df <- read.delim(here::here('data/input','crab_open_dates.txt')); head(dates_df)
```
<br>

Make sure that all dates are `Posixt` objects.
```{r}
dates_df$odate <- mdy(dates_df$odate)
fishtix$tdate <- date(parse_date_time(fishtix$tdate, orders=c("ymd", "mdy")))
```
<br>


# 2: Seasonal Participation Networks

For each crab year and port group, this code will:

1. split the data into an early and late season

2. Create a `graph` object for each fisheries participation network, and then save the object as an R data file

3. Plot the network graphs

4. Calculate network-level statistics, and save to an output data frame.


```{r eval=TRUE}
for(y in years){
  for(p in myports){
    cat(p, "\n")
    
    #### Subset data ####
    # subset dates data frame to get opening date for given port group (not specific for year)
    open_info <- filter(dates_df, pcgroup == p)
    open_month <- month(filter(dates_df, pcgroup == p)$odate); open_day <- day(filter(dates_df, pcgroup == p)$odate)
    # create open date object for the current fishtix year, 'y'
    if(year(open_info$odate) == open_info$crab_year){
      open_date <- paste0(y, "-", open_month, "-", open_day)
    } else{
      open_date <- paste0(y+1, "-", open_month, "-", open_day)
    }
    open_date <- ymd(open_date)
    # pull fish ticket data for given year, port group, before Dcrab opened
    early_dat <- fishtix %>%
      filter(crab_year == y & pcgroup == p) %>%
      filter(tdate < open_date)
    if(y==2015 & rm_crab){
      early_dat <- early_dat %>%
        filter(spid != "DCRB")
    }
    # pull fish ticket data for given year, port group, after Dcrab opened
    late_dat <- fishtix %>%
      filter(crab_year == y & pcgroup == p) %>%
      filter(tdate >= open_date)
    
    #### Read in metier names key ####
    met_names <- read.delim(here::here(indir, paste0(p, "_metier_names.txt")), stringsAsFactors = FALSE)
    
    
    ##### EARLY SEASON ####
    
    # Create igraph object
    ## using participation_network function
    early_g <- participation_network_crabyear(early_dat, filter = TRUE, filter_subgraph =FALSE, min_vessels = 3, min_contribution = 0.10)
    if(any(is.na(early_g))){
      message("\n\n WARNING: No fishing at ", p, " during early season in crab year ", y, "!\n\n")
    } else if(vcount(early_g) > 0){
      ## add in the full metier names as vertex attribute
      filtered_metier_info <- as.data.frame(x = V(early_g)$name)
      colnames(filtered_metier_info) <- "metier.abbrev"
      filtered_metier_info <- left_join(filtered_metier_info, met_names, by="metier.abbrev")
      V(early_g)$common_name <- as.character(filtered_metier_info$metier.name)
      ## save vertex colors as attribute
      vertex_cols <- vertex_color(g = early_g)
      V(early_g)$colors <- vertex_cols
      ## save objects
      rds_name <- ifelse(rm_crab,  paste0("igraph_", p, "_", y, "_early_rmCrab.rds"),  paste0("igraph_", p, "_", y, "_early.rds"))
      saveRDS(early_g, file = here::here(outdir, rds_name))
      
    # Calculate network-level statistics
    ## using the `net_stats` function
    early_stats <- net_stats(early_g, y = y, stats="network")
    ## add on period (closure/open) and port group info
    early_stats <- early_stats %>%
      mutate(period = rep("early", times=dim(early_stats)[1])) %>%
      mutate(pcgroup = rep(p, times=dim(early_stats)[1]))
    } #end 'else' close-g has > 0 vertex
    
    
    ##### LATE SEASON ####
    # Create igraph object
    ## using participation_network function
    late_g <- participation_network_crabyear(late_dat, filter = TRUE, filter_subgraph = FALSE, min_vessels = 3, min_contribution = 0.10)
    ## add in the full metier names as vertex attribute
    filtered_metier_info <- as.data.frame(x = V(late_g)$name)
    colnames(filtered_metier_info) <- "metier.abbrev"
    filtered_metier_info <- left_join(filtered_metier_info, met_names, by="metier.abbrev")
    V(late_g)$common_name <- as.character(filtered_metier_info$metier.name)
    ## save vertex colors as attribute
    vertex_cols <- vertex_color(g = late_g)
    V(late_g)$colors <- vertex_cols
    ## save objects
    rds_name <- ifelse(rm_crab,  paste0("igraph_", p, "_", y, "_late_rmCrab.rds"),  paste0("igraph_", p, "_", y, "_late.rds"))
    saveRDS(late_g, file = here::here(outdir, rds_name))

    # Calculate network-level statistics
    ## using the `net_stats` function
    late_stats <- net_stats(late_g, y = y, stats="network")
    ## add on period (closure/open) and port group info
    late_stats <- late_stats %>%
      mutate(period = rep("late", times=length(y))) %>%
      mutate(pcgroup = rep(p, times=length(y)))
    
    #### SAVE NETWORK STATS OUTPUT ####
    ## merge network stats from closure / open period
    if(length(V(early_g)$name) == 0){
      tmp_stats <- early_stats
    } else{
      tmp_stats <- rbind(early_stats, late_stats)
    }
    ## append to network stats output data frame
    if(exists('mystats')){
      mystats <- rbind(mystats, tmp_stats)
    } else{
      mystats <- tmp_stats
    }
    
  } #end (p in myports)
  cat("\nfinished with all port groups in crab year ", y, "\n-----\n")
} #end (y in years)
```
<br>

Check the network stats data frame
```{r eval=TRUE}
head(mystats)
```
<br>


Write stats out to file. 2008_2015_CA_CrabPorts_NoSubgraph_01contribution.csv
```{r eval=TRUE}
filename <- paste0("NetworkStats_", min(years), "_", max(years), "_", length(myports), "pcgroups_", ifelse(rm_crab, 'rmCrab_','_'), contr_cutoff*100, "pContribution.csv")

write.csv(mystats,here::here(statdir, filename), row.names=FALSE)
```
<br>


# 3: Seasonal Participation Network Statistics

If you have already produced a `graph` object for each fisheries participation network, this section of the script will read in those objects and calculate network-level statistics

```{r eval=FALSE}
for(y in years){
  for(p in myports){
    cat(p, "\n")
    
    ##### EARLY SEASON ####
    
    # Read in igraph object
    rds_name <- ifelse(rm_crab,  paste0("igraph_", p, "_", y, "_early_rmCrab.rds"),  paste0("igraph_", p, "_", y, "_early.rds"))
    ## (if the object exists)
    if(file_test("-f", here::here(outdir, rds_name))){
      early_g <- readRDS(here::here(outdir, rds_name))
      
      # Calculate network-level statistics
      ## using the `net_stats` function
      early_stats <- net_stats(early_g, y = y, stats="network")
      ## add on period (closure/open) and port group info
      early_stats <- early_stats %>%
        mutate(period = rep("early", times=dim(early_stats)[1])) %>%
        mutate(pcgroup = rep(p, times=dim(early_stats)[1]))
      ## save
      tmp_stats <- early_stats
    } else{message("Warning: Could not find early season network for ",p,", ", y, "\n")}
    
    
    ##### LATE SEASON ####
    
    # Read in igraph object
    rds_name <- ifelse(rm_crab,  paste0("igraph_", p, "_", y, "_late_rmCrab.rds"),  paste0("igraph_", p, "_", y, "_late.rds"))
    ## (if the object exists)
    if(file_test("-f", here::here(outdir, rds_name))){
      late_g <- readRDS(here::here(outdir, rds_name))
      
      # Calculate network-level statistics
      ## using the `net_stats` function
      late_stats <- net_stats(late_g, y = y, stats="network")
      ## add on period (closure/open) and port group info
      late_stats <- late_stats %>%
        mutate(period = rep("late", times=length(y))) %>%
        mutate(pcgroup = rep(p, times=length(y)))
    }
    
    #### SAVE NETWORK STATS OUTPUT ####
    ## merge data from early / late season
    if(!exists('early_stats')){
      tmp_stats <- late_stats
      rm(late_stats)
    } else{
      tmp_stats <- rbind(early_stats, late_stats)
      rm(early_stats, late_stats)
    }
    
    ## append to network stats output data frame
    if(exists('mystats')){
      mystats <- rbind(mystats, tmp_stats)
    } else{
      mystats <- tmp_stats
    }
    
  } #end (p in myports)
  cat("\nfinished with all port groups in crab year ", y, "\n-----\n")
} #end (y in years)
```
<br>

Check the network stats data frame
```{r eval=FALSE}
head(mystats)
```
<br>


Write stats out to file. 2008_2015_CA_CrabPorts_NoSubgraph_01contribution.csv
```{r eval=FALSE}
filename <- paste0("NetworkStats_", min(years), "_", max(years), "_", length(myports), "pcgroups_", ifelse(rm_crab, 'rmCrab_','_'), contr_cutoff*100, "pContribution.csv")

write.csv(mystats,here::here(statdir, filename), row.names=FALSE)
```
<br>

---

Script Runtime:
```{r}
Sys.time() - script_start_time
```
