==> scripts/07_explore_network_stats.Rmd <==
---
title: "Explore Network Statistics"
author: "M. Fisher"
date: "Written Oct. 25, 2019. Last Run `r Sys.Date()`"
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
  pdf_document:
    highlight: haddock
    number_sections: yes
    toc: yes
    toc_depth: '3'
geometry: margin=1in
subtitle: Preparation for network analysis in Fisher et al.
fontsize: 11pt
---


# Description

First, group fishing communities (port groups) into regions, according to an NMDS cluster analysis.

Then prepare for generalized linear models (GLMs). Check for a correlation between network size (node count) and network statistics, to determine whether network size needs to be incorporated into the GLMs. Do some pre-model visualization; graph network statistics through time and across closure durations (this will produce Figures S2-S4).

The three core network metrics are: 

1. Weighted Network Centralization (alt: unweighted)

2. Weighted Modularity (alt:unweighted)

3. Edge Density (alt: mean degree)


<br>
```{r "setup", include=FALSE}
if(!require("here")) {install.packages("here")}
library(here)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())

## start time for full script
script_start_time <- Sys.time()
```
<br>

This script requires the following packages. 
```{r packages, message=FALSE, warning=FALSE}
if(!require("tidyverse")) {install.packages("tidyverse")}
if(!require("foreign")) {install.packages("foreign")}
if(!require("lubridate")) {install.packages("lubridate")}
if(!require("vegan")) {install.packages("vegan")}
if(!require("labdsv")) {install.packages("labdsv")}
if(!require("plotrix")) {install.packages("plotrix")}
if(!require("facetscales")) {devtools::install_github("zeehio/facetscales"); library(facetscales)}
if(!require("ggplot2")) {install.packages("ggplot2")}
if(!require("ggord")) {devtools::install_github("fawda123/ggord"); library(facetscales)}
if(!require("gridExtra")) {install.packages("gridExtra")}
if(!require("cowplot")) {install.packages("cowplot")}
if(!require("PNWColors")) {devtools::install_github("jakelawler/PNWColors"); library(PNWColors)}
```
<br>

# User Inputs

Select your directories.
```{r}
## directory with network statistics 
indir <- 'results/statistics'

## directory to save figures
pngdir <- 'results/figures'
```
<br>

What is the name of the `.csv` file with the network statistics from script 06?
```{r}
filename <- "2008_2017_CA_ParticipationNetworkMetrics.csv"
```
<br>

Specify which network statistics to retain from the full data set for the NMDS, and for the GLMs. The NMDS variables should include either all or a subset of the variables that will be used in the GLMs
```{r}
nmds_vars <- c("ed","nc_weighted","m_weighted")
myvars <- c("mean_deg","ed","nc","nc_weighted","m","m_weighted")
```
<br>



# 1: Data


Network Statistics. Add in regional designation for each port group.
```{r}
stats <- read.csv(here::here(indir, filename)) %>%
  dplyr::select(pcgroup, y, period, N, all_of(myvars)) %>%
  mutate(R = ifelse(pcgroup %in% c("CCA", "ERA", "BGA", "BDA"), "north", "central"))
colnames(stats)
```
<br>


Crab fishing season delays (in days)
```{r}
delays <- read.csv(here::here('data/input', 'DCRB_Historic_Closures_CA_updated.csv'))
head(delays)
```
<br>

# 2. NMDS

The non-metric multidimensional scaling (NMDS) ordination will collapse port group information from many network metrics into a 2D visualization. Since NMDS uses rank orders, we don't have to make any assumptions about the nature of the network metrics data. The goal for the NMDS is to evaluate the regional designations of each port group, particularly Bodega Bay. Although Bodega Bay is technically within the central California management area for the commercial Dungeness crab fishery, for the purpose of some analyses it has been considered part of northern California. 

Subset the stats data for the NMDS variables, and for observations (port group / year / season combinations) which have complete data.
```{r}
nmds_stats <- stats %>%
  dplyr::select(pcgroup, y, period, all_of(nmds_vars)) %>%
  filter_at(vars(all_of(nmds_vars)), all_vars(!is.na(.)))
```
<br>

NMDS analyses will be done separately for the early and late seasons. 
```{r}
nmds_stats_early <- filter(nmds_stats, period=="early")
nmds_stats_late <- filter(nmds_stats, period=="late")
```
<br>

## 2.1 Early Season

Relativize the network metrics by range, and run the `metaMDS` function.
```{r}
nmds_stats_early.r <- decostand(nmds_stats_early[4:ncol(nmds_stats_early)], method="range",MARGIN=2)
early.nmds <- metaMDS(comm=nmds_stats_early.r,autotransform=FALSE,distance="euclidean")
```
```{r echo=FALSE}
stressplot(early.nmds)
```
**The stress is:`r round(early.nmds$stress,5)`**
<br>


## 2.2. Late Season

Relativize the network metrics by range, and run the `metaMDS` function.
```{r}
nmds_stats_late.r <- decostand(nmds_stats_late[4:ncol(nmds_stats_late)], method="range",MARGIN=2)
late.nmds <- metaMDS(comm=nmds_stats_late.r,autotransform=FALSE,distance="euclidean")
```
```{r echo=FALSE}
stressplot(late.nmds)
```
**The stress is:`r round(late.nmds$stress,5)`**
<br>

## 2.3. Fig. S1

Graph the ordinations. Note: you will get a warning message from the `ggord` function about replacing the existing scale for shape. Ignore it!

The early season, fill by port group:
```{r}
nmds_stats_early <- nmds_stats_early %>%
  ## replace codes with names
  mutate(pcgroup_name = recode(nmds_stats_early$pcgroup, 
                               CCA="Crescent City", ERA="Eureka",
                               BGA="Fort Bragg", BDA="Bodega Bay",
                               SFA="San Francisco",
                               MRA="Morro Bay",
                               MNA="Monterey")) %>%
  ## assign expected north / central, add in NMDS scores
  mutate(Region=ifelse(pcgroup_name %in% c("Crescent City","Eureka","Fort Bragg","Bodega Bay"), "North","Central"),
         port_group=ifelse(pcgroup_name=="Bodega Bay","Bodega Bay","Other"),
         one = early.nmds$points[,1,drop=TRUE], 
         two = early.nmds$points[,2,drop=TRUE])

## plot ordination
e.ordp <- ggord(ord_in=early.nmds, grp_in=nmds_stats_early$pcgroup_name, ellipse_pro=c(0.10,0.90),
      cols=pnw_palette(name="Sunset",n=7,type="discrete"),arrow=NULL,vec_lab="") +
  scale_shape_manual('Groups', values = c(16,16,16,17,16,16,16)) +
  theme(panel.grid.major=element_blank(),panel.grid.minor=element_blank(),
        axis.text=element_blank(),legend.title = element_blank(), legend.text=element_text(size=16),
        legend.position = "bottom")
```
<br>

The early season, fill by region:
```{r}
nmds_stats_early$port_group <- factor(nmds_stats_early$port_group,levels=c("Other","Bodega Bay"))

# create ggord and remove points layer
e.ord <- ggord(ord_in=early.nmds, grp_in=nmds_stats_early$Region, ellipse_pro=c(0.10,0.90), cols=c("#41476b","#de9b71"), arrow=NULL,vec_lab="")
e.ord$layers[[1]] <- NULL

# new point scales
e.ordr <- e.ord + geom_point(data = nmds_stats_early, aes(shape = port_group, colour = Region, group = Region), size= 4, alpha = 0.7) + 
  scale_shape_manual(values = c(1, 17)) + 
  guides(fill = guide_legend(override.aes= list(shape = NA))) + 
  theme(legend.title = element_blank(), legend.text=element_text(size=16),
        legend.position = "bottom",
        panel.grid.major=element_blank(),panel.grid.minor=element_blank(),
        axis.text=element_blank(), axis.ticks=element_blank())
```
<br>

The late season, fill by port group:
```{r}
nmds_stats_late <- nmds_stats_late %>%
  ## replace codes with names
  mutate(pcgroup_name = recode(nmds_stats_late$pcgroup, 
                               CCA="Crescent City", ERA="Eureka",
                               BGA="Fort Bragg", BDA="Bodega Bay",
                               SFA="San Francisco",
                               MRA="Morro Bay",
                               MNA="Monterey")) %>%
  ## assign expected north / central, add in NMDS scores
  mutate(Region=ifelse(pcgroup_name %in% c("Crescent City","Eureka","Fort Bragg","Bodega Bay"), "North","Central"),
         port_group=ifelse(pcgroup_name=="Bodega Bay","Bodega Bay","Other"),
         one = late.nmds$points[,1,drop=TRUE], 
         two = late.nmds$points[,2,drop=TRUE])

## plot ordination
l.ordp <- ggord(ord_in=late.nmds, grp_in=nmds_stats_late$pcgroup_name, ellipse_pro=c(0.10,0.90),
      cols=pnw_palette(name="Sunset",n=7,type="discrete"),arrow=NULL,vec_lab="",alpha_el=0.45) +
  scale_shape_manual('Groups', values = c(16,16,16,17,16,16,16)) +
  theme(panel.grid.major=element_blank(),panel.grid.minor=element_blank(),
        axis.text=element_blank(),legend.position="none")
```
<br>

The late season, fill by region:
```{r}
nmds_stats_late$port_group <- factor(nmds_stats_late$port_group,levels=c("Other","Bodega Bay"))

# create ggord and remove points layer
l.ord <- ggord(ord_in=late.nmds, grp_in=nmds_stats_late$Region, ellipse_pro=c(0.10,0.90), cols=c("#41476b","#de9b71"), arrow=NULL,vec_lab="")
l.ord$layers[[1]] <- NULL

# new point scales
l.ordr <- l.ord + geom_point(data = nmds_stats_late, aes(shape = port_group, colour = Region, group = Region), size= 4, alpha = 0.7) + 
  scale_shape_manual(values = c(1, 17)) + 
  guides(fill = guide_legend(override.aes= list(shape = NA))) + 
  theme(legend.position="none",
        panel.grid.major=element_blank(),panel.grid.minor=element_blank(),
        axis.text=element_blank(), axis.ticks=element_blank())
```


Figure S1. 
```{r}
plot_grid(e.ordp,l.ordp,nrow=2,rel_heights=c(0.77,1),labels=c("a","c"),label_size=18)
plot_grid(e.ordr,l.ordr,nrow=2,labels=c("a","c"),rel_heights=c(1,0.69))

png(here::here("results/figures","FigureS1_ac.png"),height=480,width=500)
print(plot_grid(e.ordp,l.ordp,nrow=2,rel_heights=c(1,0.74),labels=c("a","c"),label_size=18))
dev.off()

png(here::here("results/figures","FigureS1_bd.png"),height=480,width=500)
print(plot_grid(e.ordr,l.ordr,nrow=2,labels=c("b","d"),rel_heights=c(1,0.77),label_size=18))
dev.off()
```
<br>


# 3: Spearman Rank Correlation Test

This function runs the correlation test on the data for each region within each season. It takes a data frame with 4 columns: period, region, network size (N) and the statistics of interest.

```{r}
size_spearman <- function(mydata, metric){
  output <- data.frame(statistic = as.character(),
                       period = as.character(),
                       region = as.character(),
                       sp.rho = as.numeric(),
                       p.value = as.numeric())
  for(p in c("early","late")){
    for(r in unique(mydata$R)){
      tmpdat <- filter(mydata, R==r & period==p)
      my.cor <- cor.test(tmpdat[,3], tmpdat[,4], method=("spearman"))
      output <- rbind(output,
                      data.frame(statistic=metric,period=p, region=r,sp.rho=my.cor$estimate,p.value=my.cor$p.value))
    }
  }
  return(output)
}
```
<br>

Save an empty data frame
```{r}
spearman_dat <- data.frame(statistic = as.character(),
                           period = as.character(),
                           region = as.character(),
                           sp.rho = as.numeric(),
                           p.value = as.numeric())
```
<br>

Run Spearman Rank for each metric and append the output to the empty data frame.
```{r}
for(v in myvars){
  tmpstats <- stats %>%
    dplyr::select(period, R, N, all_of(v))
  tmp_out <- size_spearman(mydata=tmpstats,metric=v)
  spearman_dat <- rbind(spearman_dat,
                        tmp_out)
}
```
<br>

Write out test results
```{r}
testfile <- paste0(str_split(filename,'\\.')[[1]][1], "_SPEARMAN.csv")
write.csv(spearman_dat, here::here(indir,testfile))
```
<br>


# 4: Graph Network Statistics

This section has hard-coded the network metrics for graphing; if you want to use any metrics not included in Fisher et al., you will need to manually edit the following code.

## 4.1. Figure S2

Calculate the yearly average for each metric, for each region / season. Fix the metric names.
```{r}
yearly_avg <- stats %>%
  pivot_longer(cols=colnames(stats)[which(!(colnames(stats) %in% c("y","period","pcgroup","R")))]) %>%
  filter(name != "N") %>%
  group_by(y,period,R,name) %>%
  summarise(avg=mean(value,na.rm=TRUE),sdev=std.error(value,na.rm=TRUE))


yearly_avg$name <- as.character(yearly_avg$name)
yearly_avg$name[which(yearly_avg$name=="ed")] <- "Edge Density"
yearly_avg$name[which(yearly_avg$name=="mean_deg")] <- "Mean Degree"
yearly_avg$name[which(yearly_avg$name=="nc")] <- "Centralization\n (UnW)"
yearly_avg$name[which(yearly_avg$name=="nc_weighted")] <- "Centralization"
yearly_avg$name[which(yearly_avg$name=="m")] <- "Modularity\n (UnW)"
yearly_avg$name[which(yearly_avg$name=="m_weighted")] <- "Modularity"
yearly_avg$name <- factor(yearly_avg$name, levels=rev(c("Edge Density","Mean Degree",
                                                            "Centralization","Centralization\n (UnW)",
                                                            "Modularity","Modularity\n (UnW)")))

```
<br>

Set y axis limits and breaks
```{r}
scales_y <- list(
  `Edge Density` = scale_y_continuous(limits = c(0.2,0.8)),
  `Mean Degree` = scale_y_continuous(limits = c(0,8)),
  `Centralization` = scale_y_continuous(limits = c(0,0.7)),
  `Centralization\n (UnW)` = scale_y_continuous(limits=c(0.2,1), breaks=seq(0.2,1,by=0.2)),
  `Modularity` = scale_y_continuous(limits=c(-0.4,0.6),breaks=seq(-0.4,0.6,by=0.2)),
  `Modularity\n (UnW)` = scale_y_continuous(limits=c(-0.5,0.3))
)
```
<br>

Separate plots for early and late season
```{r}
myplot.e <- ggplot(data=filter(yearly_avg, period=="early"), aes(y, fill=R, lty=R)) +
  geom_vline(aes(xintercept=2015), size=10,color="grey67") +
  geom_ribbon(aes(ymin=avg-sdev,ymax=avg+sdev),alpha=0.5) +
  geom_line(aes(y=avg,col=R)) +
  geom_point(aes(y=avg,col=R)) +
  facet_grid_sc(rows = vars(name), scales = list(y = scales_y),as.table=FALSE) +
  scale_fill_manual(values=c("cadetblue1","darkseagreen2"), name="Region") +
  scale_color_manual(values=c("dodgerblue3 ","aquamarine4"), name="Region") +
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) +
  ylab("") +
  xlab("Crab Year") +
  theme_classic() + theme(axis.text.x=element_text(angle=90,hjust=1,vjust=0.5),
                          axis.title=element_text(size=13),
                          legend.position="none",
                          strip.background = element_blank(),
                          strip.text = element_blank(),
                          panel.border = element_rect(colour = "black", fill=NA, size=0.25),
                          axis.line=element_blank())

myplot.l <- ggplot(data=filter(yearly_avg, period=="late"), aes(y, fill=R, lty=R)) +
  geom_vline(aes(xintercept=2015), size=10,color="grey67") +
  geom_ribbon(aes(ymin=avg-sdev,ymax=avg+sdev),alpha=0.5) +
  geom_line(aes(y=avg,col=R),lwd=0.5) +
  geom_point(aes(y=avg,col=R)) +
  facet_grid_sc(rows = vars(name), scales = list(y = scales_y),as.table=FALSE) +
  scale_fill_manual(values=c("cadetblue1","darkseagreen2"), name="Region") +
  scale_color_manual(values=c("dodgerblue3 ","aquamarine4"), name="Region") +
  scale_linetype_discrete(name="Region") +
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) +
  ylab("") +
  xlab("Crab Year") +
  theme_bw() + theme(axis.text.x=element_text(angle=90,hjust=1,vjust=0.5),
                     axis.title.x=element_text(size=13),
                     axis.text.y=element_blank(),
                     axis.title.y=element_blank(),
                     strip.text=element_text(size=13),
                     legend.title=element_text(size=13),legend.text=element_text(size=13),
                     panel.grid.major=element_blank(),
                     panel.grid.minor=element_blank())
```
<br>

Save figure
```{r}
png(here::here(pngdir,'FigureS2.png'),
    width = 6.7, height = 8, units = 'in', res = 300)
plot_grid(myplot.e,myplot.l,rel_widths = c(0.9,1), labels=c("(a)","(b)"))
dev.off()
```
<br>

## 4.2. Figure S3

Pre-shock averages of each metric during the early and late seasons.
```{r}
preshock_avg <- stats %>%
  filter(y %in% c(2012,2013,2014)) %>%
  pivot_longer(cols=colnames(stats)[which(!(colnames(stats) %in% c("y","period","pcgroup","R")))]) %>%
  filter(name != "N") %>%
  group_by(period,R,name) %>%
  summarise(avg=mean(value,na.rm=TRUE),sdev=std.error(value,na.rm=TRUE))


preshock_avg$name <- as.character(preshock_avg$name)
preshock_avg$name[which(preshock_avg$name=="ed")] <- "Edge Density"
preshock_avg$name[which(preshock_avg$name=="mean_deg")] <- "Mean Degree"
preshock_avg$name[which(preshock_avg$name=="nc")] <- "Centralization\n (UnW)"
preshock_avg$name[which(preshock_avg$name=="nc_weighted")] <- "Centralization"
preshock_avg$name[which(preshock_avg$name=="m")] <- "Modularity\n (UnW)"
preshock_avg$name[which(preshock_avg$name=="m_weighted")] <- "Modularity"
preshock_avg$name <- factor(preshock_avg$name, levels=c("Edge Density","Mean Degree","Centralization","Centralization\n (UnW)",
                                                                    "Modularity", "Modularity\n (UnW)"))
preshock_avg$period <- recode(preshock_avg$period, closure="Early",open="Late")
```


Set the limits and breaks for the y axis.
```{r}
scales_y <- list(
  `Edge Density` = scale_y_continuous(limits = c(0.2,0.8)),
  `Mean Degree` = scale_y_continuous(limits = c(0,6)),
  `Centralization` = scale_y_continuous(limits = c(0.1,0.7)),
  `Centralization\n (UnW)` = scale_y_continuous(limits=c(0.2,0.9), breaks=seq(0.2,0.9,by=0.2)),
  `Modularity` = scale_y_continuous(limits=c(-0.4,0.6),breaks=seq(-0.4,0.6,by=0.2)),
  `Modularity\n (UnW)` = scale_y_continuous(limits=c(-0.5,0.3))
)
```
<br>

plot
```{r}
myplot1 <- ggplot(data=preshock_avg, aes(x=period,y=avg)) +
  geom_point(aes(col=R),size=2) +
  geom_path(aes(col=R,group=R),size=0.5) +
  geom_errorbar(aes(ymin=avg-sdev,ymax=avg+sdev,col=R,group=R), width=0.2) +
  facet_grid_sc(rows = vars(name),scales=list(y = scales_y)) +
  scale_color_manual(values=c("dodgerblue3","aquamarine4"), name="") +
  ylab("") +
  xlab("Season") +
  theme_classic() +
  theme(axis.text.x=element_text(size=14),
        axis.text.y=element_text(size=14,vjust=0.6),
        axis.title=element_text(size=16),
        strip.text=element_text(size=13.5),
        legend.text=element_text(size=14),
        panel.border = element_rect(colour = "black", fill=NA, size=0.25),
        axis.line=element_blank(),
        plot.margin = unit(c(0, 0, 0, 0), "cm"))
```


Save 
```{r}
png(here::here(pngdir, "FigureS3.png"),
    width = 4, height = 8, units = 'in', res = 300)
myplot1
dev.off()
```
<br>


## 4.3. Figure S4

Match the closure duration data with the network metrics data. Then categorize the closure duration, into no closure (0 days; "none"), a mid-duration closure (< 50 days; "medium"), and a high duration closure (> 50 days; "high").
```{r}
stats_expand <- left_join(stats, delays, by=c("y","pcgroup")) %>%
  mutate(closure.duration = ifelse(days.closed == 0, "none",
                                   ifelse(days.closed < 50, "medium", "high")))
```
<br>


Averages of each metric across closure duration (none / medium / high) between 2008 and 2017 crab years.
```{r}
duration_avg <- stats_expand %>%
  pivot_longer(cols=colnames(stats)[which(!(colnames(stats) %in% c("y","period","pcgroup","R")))]) %>%
  filter(name != "N") %>%
  group_by(period,R,name, closure.duration) %>%
  summarise(avg=mean(value,na.rm=TRUE),sdev=std.error(value,na.rm=TRUE))


duration_avg$name <- as.character(duration_avg$name)
duration_avg$name[which(duration_avg$name=="ed")] <- "Edge Density"
duration_avg$name[which(duration_avg$name=="mean_deg")] <- "Mean Degree"
duration_avg$name[which(duration_avg$name=="nc")] <- "Centralization\n (UnW)"
duration_avg$name[which(duration_avg$name=="nc_weighted")] <- "Centralization"
duration_avg$name[which(duration_avg$name=="m")] <- "Modularity\n (UnW)"
duration_avg$name[which(duration_avg$name=="m_weighted")] <- "Modularity"
duration_avg$name <- factor(duration_avg$name, levels=c("Edge Density","Mean Degree","Centralization","Centralization\n (UnW)",
                                                                    "Modularity", "Modularity\n (UnW)"))
duration_avg$period <- recode(duration_avg$period, closure="early",open="late")
duration_avg$R <- recode(duration_avg$R, central="Central",north="North")
duration_avg$closure.duration <- factor(duration_avg$closure.duration, levels=c("none","medium","high"))
```
<br>

Set the limits and breaks for the y axis.
```{r}
scales_y <- list(
  `Edge Density` = scale_y_continuous(limits = c(0.2,0.7)),
  `Mean Degree` = scale_y_continuous(limits = c(0,6)),
  `Centralization` = scale_y_continuous(limits = c(0.2,0.6),breaks=seq(0.2,0.6,by=0.2)),
  `Centralization\n (UnW)` = scale_y_continuous(limits=c(0.1,0.9), breaks=seq(0.1,0.9,by=0.2)),
  `Modularity` = scale_y_continuous(limits=c(-0.4,0.6),breaks=seq(-0.4,0.6,by=0.2)),
  `Modularity\n (UnW)` = scale_y_continuous(limits=c(-0.4,0.2))
)
```
<br>

plot the early season
```{r}
myplot.e <- ggplot(data=filter(duration_avg, period=="early"), aes(x=closure.duration,y=avg)) +
  geom_point(aes(col=R),size=2) +
  geom_path(aes(col=R,group=R),size=0.5) +
  geom_errorbar(aes(ymin=avg-sdev,ymax=avg+sdev,col=R,group=R), width=0.2) +
  facet_grid_sc(rows = vars(name), scales=list(y = scales_y)) +
  scale_color_manual(values=c("dodgerblue3","aquamarine4"), name="") +
  ylab("") +
  xlab("Closure Duration") +
  theme_classic() +
  theme(axis.text.x=element_text(size=14,angle=90,hjust=1,vjust=0.5),
        axis.text.y=element_text(size=14,vjust=0.6),
        axis.title=element_text(size=16),
        strip.text=element_blank(),
        legend.position="none",
        panel.border = element_rect(colour = "black", fill=NA, size=0.25),
        axis.line=element_blank())
myplot.e
```
<br>

plot the late season
```{r}
myplot.l <- ggplot(data=filter(duration_avg, period=="late"), aes(x=closure.duration,y=avg)) +
  geom_point(aes(col=R),size=2) +
  geom_path(aes(col=R,group=R),size=0.5) +
  geom_errorbar(aes(ymin=avg-sdev,ymax=avg+sdev,col=R,group=R), width=0.2) +
  facet_grid_sc(rows = vars(name), scales=list(y = scales_y)) +
  scale_color_manual(values=c("dodgerblue3","aquamarine4"), name="Region") +
  ylab("") +
  xlab("Closure Duration") +
  theme_classic() +
  theme(axis.text.x=element_text(size=14,angle=90,hjust=1,vjust=0.5),
        axis.text.y=element_blank(),
        axis.title.x=element_text(size=16),
        axis.title.y=element_blank(),
        strip.text=element_text(size=13.5),
        strip.background = element_rect(fill="grey88"),
        legend.text=element_text(size=14),
        legend.title=element_text(size=14),
        panel.border = element_rect(colour = "black", fill=NA, size=0.25),
        axis.line=element_blank())
myplot.l
```
<br>

Save figure
```{r}
png(here::here(pngdir,'FigureS4.png'),
    width = 7, height = 9, units = 'in', res = 300)
plot_grid(myplot.e,myplot.l,rel_widths = c(0.8,1), labels=c("(a)","(b)"))
dev.off()
```
<br>
