==> scripts/09b_directed_vessel_network_late.Rmd <==
---
title: "Directed D. crab Vessel Networks, Late Season"
author: "M. Fisher"
date: "Written August 5, 2019. Last Run `r Sys.Date()`"
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
  pdf_document:
    highlight: haddock
    number_sections: yes
    toc: yes
    toc_depth: '3'
geometry: margin=1in
subtitle: Preparation for network analysis in Fisher et al.
fontsize: 11pt
---

# Description

Create directed networks to show flow of Dungeness crab vessels between fisheries in the late seasons of the 2015 --> 2016 crab years. If the vessel continued participation in the same fishery during both late seasons, fishing activity is showed using a self-loop instead of a directed edge. 

Nodes represent fishing strategies, sized according to the number of participating vessels in the 2016 late season. This includes (1) Dungeness & non-Dungeness crab fisheries, (2) movement to land fish at a different port group, (3) a pause in participation in any fishery in California.


<br>
```{r "setup", include=FALSE}
if(!require("here")) {install.packages("here")}
library(here)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())

## start time for full script
script_start_time <- Sys.time()
```
<br>

This script requires the following packages. 
```{r packages, message=FALSE, warning=FALSE}
if(!require("tidyverse")) {install.packages("tidyverse")}
if(!require("lubridate")) {install.packages("lubridate")}
if(!require("igraph")) {install.packages("igraph")}
if(!require("ggplot2")) {install.packages("ggplot2")}
if(!require("grid")) {install.packages("grid")}
if(!require("gridExtra")) {install.packages("gridExtra")}
```
<br>

And calls the following functions:
```{r}
source(here::here("R","directed_network_matrix_late.R"))
source(here::here("R","directed_network_collapse_confidential_late.R"))
source(here::here("R","collapse_confidential_matrix_rows.R"))
source(here::here("R","fill_adjacency_matrix.R"))
source(here::here("R","color_network_nodes.R"))
```
<br>

# User Inputs 

Select your directories.
```{r get_dir}
## location of fish ticket with metiers assigned, metier key for each port group
indir <- 'data/input/metiers'

## location of raw fish tickets without metiers
indir_raw <- 'data/input'

## location of vessel length .csv files
indir_vl <- 'data/input/length'

## output directory for adjacency matrix and igraph object
outdir <- 'data/networks/vessel_flow'
```
<br>

Identify the port groups that you would like to produce networks for. The port groups vector should include only those port groups which are present in the single data file produced with script 05.
```{r}
myports <- c("CCA", "ERA", "BGA", "BDA", "SFA", "MRA", "MNA")
```
<br>

Include self-loops when vessels were active in a given fishery in both the 2015 and 2016 crab years?
```{r}
self = TRUE
```
<br>

Do you want to save the confidential adjacency matrix? This will include all fishing activity, including fisheries with < 3 participating vessels. **If you are going to collapse confidential nodes into a non-confidential network object, you need to save the confidential adjacency matrix**
```{r}
save_con=TRUE
```
<br>

Set these objects according to your user inputs from Scripts 1 and 3
```{r}
## the value of `k` for the k-nearest neighbor function
k <- 1

## whether out-of-season Dungeness crab landings were removed
rm_crab = TRUE
```
<br>


# 1: Prepare Network Data


# 1: Prepare Network Data

Read in the landings data, from the file containing all fish tickets across port groups, within a single crab year.
```{r}
years <- c(2014,2015)

for(y in years){
  tmptix <- read.csv(here::here(indir, paste0("FishTix_crab", y, "_metiers_k1_recoded.csv")), stringsAsFactors = FALSE)
  if(exists('fishtix')){
    fishtix <- rbind(fishtix, tmptix)
  } else{
    fishtix <- tmptix
  }
}

rm(tmptix)
colnames(fishtix)
unique(fishtix$crab_year)
```
<br>

Although the metier analysis and participation networks focus on crab ports, for the purpose of tracking vessel movement we also want to know whether vessels landed any catch at non-crab ports. For manuscript, only other California port groups were included, but this step could incorporate Oregon / Washington port groups as well. 
```{r}
noncrab_tix <- read.csv(here::here(indir_raw, paste0("fish_tickets_crab",years[2],"_processed_for_networks.csv"))) %>%
  filter(agid=="C")
```
<br>

Read in the text file with opening dates for each port group.
```{r}
dates_df <- read.delim(here::here('data/input', 'crab_open_dates.txt')); head(dates_df)
```
<br>

The vessel length key for 2015-16 registered fishing vessels will be used to assign vessel lengths.  
```{r}
length_key = read.csv(here::here(indir_vl, paste0('dcrb_vessel_length_key_crab',years[2],'.csv')))

## what proportion of vessels do not have a calculated length?
prop.table(with(length_key,table(vessel_cat))); sum(is.na(length_key$vessel_cat))/length(length_key$vessel_cat)
```
<br>

Any vessels which landed Dungeness crab in 2015-16 but do not have a length for 2015-16, will be assigned a vessel length from the 2014-15 data.
```{r}
length_key2 = read.csv(here::here('data/input/length', paste0('dcrb_vessel_length_key_crab',years[1],'.csv'))) %>%
  filter(!(drvid %in% length_key$drvid))

length_key <- rbind(length_key,length_key2)
```
<br>

Make sure that all dates are Posixt objects.
```{r}
dates_df$odate <- mdy(dates_df$odate)
fishtix$tdate <- date(parse_date_time(fishtix$tdate, orders=c("ymd", "mdy")))
noncrab_tix$tdate <- date(parse_date_time(noncrab_tix$tdate, orders=c("ymd", "mdy")))
```
<br>

Then add the vessel lengths to the fish ticket data. What proportion of vessels in the fish ticket data are not assigned a length? These will be left out of the analyses based on vessel size categories.
```{r}
fishtix <- left_join(fishtix,dplyr::select(length_key,drvid,mean_length,vessel_cat),by=c("drvid"))
sum(is.na(fishtix$vessel_cat))/length(fishtix$vessel_cat)
```
<br>

Save fish ticket data for large (>= 40 feet in length) and small (< 40 feet in length) vessels to separate data frames.
```{r}
fishtix_lg <- filter(fishtix,vessel_cat=="large")
fishtix_sm <- filter(fishtix,vessel_cat=="small")
```
<br>


# 2: Construct Networks

For each port group, and each vessel size category:

1. create an adjacency matrix from the fish ticket data

2. collapse / filter out fisheries that are confidential (< 3 vessel participating)

3. construct a basic network from the adjacency matrix, with fishery-specific vertex size and color

4. customize the vertex names in the graph, and save the final graph object

```{r networks}
for(p in myports){
  cat("\n\nworking on port group: ",p,".\n")
  
  
  ## CREATE ADJACENCY MATRIX ##
  A_info <- gen_adj_matrix_late(fishtix_lg,p=p,self=self,write=save_con,outdir=outdir,size="large",noncrab_fishtix=noncrab_tix)
  A_lg <- A_info[[1]]; nv_lg <- A_info[[2]]
  # how many vessels?
  cat("Number large vessels at ",p,": ",nv_lg,".\n")
  
  A_info_sm <- gen_adj_matrix_late(fishtix_sm,p=p,self=self,write=save_con,outdir=outdir,size="small",noncrab_fishtix=noncrab_tix)
  A_sm <- A_info_sm[[1]]; nv_sm <- A_info_sm[[2]]
  # how many vessels?
  cat("Number small vessels at ",p,": ",nv_sm,".\n")
  
  
  ## Graph ##
  i=1
  
  for(a in list(A_lg,A_sm)){
    A <- a
    print(A)
    
    # read in the confidential vessel matrix for the collapse_confidential_late function
    if(i==1){
      vmat <- here::here(outdir,paste0(p,"_late_Amatrix_large.csv"),row.names=1)
    } else{
      vmat <- here::here(outdir,paste0(p,"_late_Amatrix_small.csv"),row.names=1)
    }
    
    # collapse confidential movement out of d.crab
    newdat <- collapse_confidential_late(A=A,vessel_matrix=as.matrix(vmat), default_other=2)   
    
    A <- newdat[[1]]
    vpf_2014 <- newdat[[2]]
    vpf_2015 <- newdat[[3]]
    
    # remove remaining confidential data
    for(j in seq(1,dim(A)[2])){
      A[ A[ , j ] < 3 , j ] <- 0
    }
    print(A)
    
    # write out
    if(i==1){
      write.csv(A, here::here(outdir, paste0(p,"_late_Amatrix_noncon_lg.csv")))
    } else{
      write.csv(A, here::here(outdir, paste0(p,"_late_Amatrix_noncon_sm.csv")))
    }
    
    
    ## CONSTRUCT BASIC NETWORK ##
    g <- graph_from_adjacency_matrix(adjmatrix = A, mode = 'directed', weighted = TRUE,diag= TRUE)
    
    # Add node class - size of node is the number of vessels participating in 2015
    ## for centered dcrb matrix, no_fishing and other_port counts are reported by the function.
    vessels_per_fishery <- as.matrix(vpf_2015)
    vessels_per_fishery <- vessels_per_fishery[V(g)$name,]
    V(g)$size <- vessels_per_fishery
    # Add node class - size of node is number of vessels participating in 2014 (plus other port / no fishing)
    vpf14 <- as.matrix(vpf_2014)
    out_row <- as.matrix(c(vessels_per_fishery["no_fishing"],vessels_per_fishery[ "other_port"]),nrow=2,ncol=1); rownames(out_row) <- c("no_fishing", "other_port")
    vpf14 <- rbind(vpf14, out_row)
    vpf14 <- vpf14[V(g)$name,]
    V(g)$size14 <- vpf14
    
    # Filter - some of this is deprecated from old collapse_confidential function
    rm_2014 <- names(which(vpf_2014 < 3))
    rm_2015 <- names(which(vpf_2015 < 3))
    metiers_rm <- rm_2014[which(rm_2014 %in% rm_2015)]
    metiers_rm <- metiers_rm[which(metiers_rm != "OTHR")]
    metiers_keep <- which(!(V(g)$name %in% metiers_rm))
    sub_g <- induced_subgraph(g, metiers_keep)
    
    # removing metiers may have removed connections to 'no fishing' or 'other port' nodes?
    metiers_rm <- c()
    if(strength(g,"no_fishing") == 0){
      metiers_rm <- c(metiers_rm, "no_fishing")
    } 
    if(strength(g,"other_port") == 0){
      metiers_rm <- c(metiers_rm, "other_port")
    }
    metiers_keep <- which(!(V(sub_g)$name %in% metiers_rm))
    sub_g <- induced_subgraph(sub_g, metiers_keep)
    
    
    # Colors for graph
    V(sub_g)$color <- vertex_color(sub_g)
    
    # Rename graph
    if(i==1){
      sub_g_lg <- sub_g
    } else{
      sub_g_sm <- sub_g
    }
    
    i=i+1
  }
  cat("created graphs for ",p,".\n")
  
  ## CUSTOMIZE NETWORK ##
  # names
  common_names = read.delim(here::here(indir, paste0(p,"_metier_names_adj.txt")))
  add_names <- data.frame(metier.num=NA,
                          metier.abbrev=c("no_fishing","other_port","OTHR_POT_HKL","OTHR"),
                          metier.spid=NA,
                          metier.gear=c(NA,NA,"POT",NA),
                          metier.name=c("No Fishing","Other Port", "Misc. Pot/H&L", "Misc. Fisheries"))
  common_names <- rbind(common_names, add_names)
  ##large
  graph_common_names <- common_names[match(V(sub_g_lg)$name,common_names$metier.abbrev),]
  V(sub_g_lg)$common_name <- as.character(graph_common_names$metier.name)
  sub_g_lg <- delete.vertices(sub_g_lg, degree(sub_g_lg)==0) #to remove self-loops, use simplify(sub_g_lg)
  if(self==FALSE){
    sub_g_lg  <- simplify(sub_g_lg)
  }
  ##small
  graph_common_names <- common_names[match(V(sub_g_sm)$name,common_names$metier.abbrev),]
  V(sub_g_sm)$common_name <- as.character(graph_common_names$metier.name)
  sub_g_sm <- delete.vertices(sub_g_sm, degree(sub_g_sm)==0)
  if(self==FALSE){
    sub_g_sm  <- simplify(sub_g_sm)
  }
  
  
  saveRDS(sub_g_sm, here::here(outdir, paste0(p,"_late_small.rds")))
  saveRDS(sub_g_lg, here::here(outdir, paste0(p,"_late_large.rds")))
  
  cat("finished with ", p, ".\n")
  

}
```
<br>




---

Script Runtime: 
```{r echo=FALSE}
Sys.time() - script_start_time
```
<br>
<br>

*Code from v8 original script. Updates on 10/30 correspond to v9 original script.*
